% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsee.R
\name{dbConnect,PgInseeDriver-method}
\alias{dbConnect,PgInseeDriver-method}
\title{Connexion au serveur Postgres}
\usage{
\S4method{dbConnect}{PgInseeDriver}(drv, ...)
}
\arguments{
\item{drv}{Un objet de type driver \code{RPostgres PqDriver}}

\item{...}{d'autres paramètres concernant la base de données sur laquelle se connecter. Consulter la documentation sur [DBI::dbConnect()]}
}
\value{
Un objet de la classe `PgInseeConnection` permettant de communiquer avec les bases de données Postgres à l'Insee.
}
\description{
Se connecter au serveur Postgres, en visualisant la structure de la base de données dans l'onglet Connections de RStudio
}
\details{
Cette méthode surcharge le comportement de la méthode parente de la classe \code{PqConnection}
}
