% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsee.R
\name{dbDisconnect,PgInseeConnection-method}
\alias{dbDisconnect,PgInseeConnection-method}
\title{Déconnexion du serveur Postgres}
\usage{
\S4method{dbDisconnect}{PgInseeConnection}(conn, ...)
}
\arguments{
\item{conn}{Un objet de type PgInseeConnection}

\item{...}{D'autres paramètres à transmettre. Consulter la documentation sur [DBI::dbDisconnect()]}
}
\value{
None - affiche un simple message de déconnexion.
}
\description{
Permet de se déconnecter du serveur - recommandé
}
\details{
Cette méthode surcharge le comportement par défaut de la classe parente \code{PqConnection}.
}
