% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgInsee.R
\name{dbConnect,PgInseeDriver-method}
\alias{dbConnect,PgInseeDriver-method}
\title{Connexion au serveur Postgres}
\usage{
\S4method{dbConnect}{PgInseeDriver}(
  drv,
  host = NULL,
  port = NULL,
  dbname = NULL,
  user = system("whoami", intern = TRUE),
  ...
)
}
\arguments{
\item{drv}{Un objet de type driver \code{RPostgres PqDriver} - par défaut \code{RPostgres::Postgres()}}

\item{host}{L'URL du serveur Postgres}

\item{port}{Le port pour se connecter au serveur Postgres}

\item{dbname}{Le nom de la base de données}

\item{user}{Par défaut, l'Idep de l'utilisateur}

\item{...}{D'autres paramètres concernant la base de données sur laquelle se connecter. Consulter la documentation sur [DBI::dbConnect()]}
}
\value{
Un objet de la classe `PgInseeConnection` permettant de communiquer avec les bases de données Postgres à l'Insee.
}
\description{
Se connecter au serveur Postgres, en visualisant la structure de la base de données dans l'onglet Connections de RStudio
}
\details{
Cette méthode surcharge le comportement de la méthode parente de la classe \code{PqConnection}
}
